requirements.txt
# Core dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1

# AI/ML frameworks
torch==2.1.2
transformers==4.36.2
numpy==1.26.3
scipy==1.11.4
scikit-learn==1.4.0
pandas==2.1.4

# Privacy & Security
phe==1.5.0
cryptography==41.0.7
pycryptodome==3.19.0

# Networking
aiohttp==3.9.1
websockets==12.0
asyncio==3.4.3

# Blockchain
web3==6.13.0
eth-account==0.10.0

# Data processing
msgpack==1.0.7
lz4==4.3.3
zstandard==0.22.0

# Monitoring & Analytics
prometheus-client==0.19.0
psutil==5.9.7

# Web framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5

# Distributed computing
ray==2.9.0
dask==2024.1.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Code quality
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Utilities
tqdm==4.66.1
rich==13.7.0
typer==0.9.0

# Production API
fastapi==0.109.0
uvicorn[standard]==0.27.0
pyjwt==2.8.0
python-multipart==0.0.6

# MLflow tracking
mlflow==2.9.2
sqlalchemy==2.0.25

# Monitoring
psutil==5.9.7
